refactor logging crate and add observer
