+++ crates/engine/src/receiver.rs
@@ -23,7 +23,7 @@ use crate::delta::{Op, Progress, apply_delta};
 use crate::io::{io_context, is_device, preallocate};
 use crate::{EngineError, ReadSeek, Result, SyncOptions, ensure_max_alloc, last_good_block};
 use checksums::ChecksumConfigBuilder;
-use logging::{NopProgressSink, ProgressSink};
+use logging::{NopObserver, Observer};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum ReceiverState {
@@ -40,7 +40,7 @@ pub struct Receiver {
     delayed: Vec<(PathBuf, PathBuf, PathBuf)>,
     #[cfg(unix)]
     link_map: meta::HardLinks,
-    progress_sink: Arc<dyn ProgressSink>,
+    progress_sink: Arc<dyn Observer>,
 }
 
 impl Default for Receiver {
@@ -59,11 +59,11 @@ impl Receiver {
             delayed: Vec::new(),
             #[cfg(unix)]
             link_map: meta::HardLinks::default(),
-            progress_sink: Arc::new(NopProgressSink),
+            progress_sink: Arc::new(NopObserver),
         }
     }
 
-    pub fn set_progress_sink(&mut self, sink: Arc<dyn ProgressSink>) {
+    pub fn set_progress_sink(&mut self, sink: Arc<dyn Observer>) {
         self.progress_sink = sink;
     }
